requirements.txt
requests==2.28.2
google-api-python-client==2.74.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
PyDrive2==1.15.0
pandas==1.5.3
gspread==5.7.2
uiautomator2==2.16.22
